+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Feat: Add `graphql-java` instrumentation (#1777)
+* Feat: Add datasource tracing with P6Spy (#1784)
+* Fix: ActivityFramesTracker does not throw if Activity has not been added (#1782)
 
 ## 5.2.4